+ Separated all engine files into .hpp and .cpp files in order to allow compilation to be done on a file-by-file basis 	! This speeds up recompilations and incidentally helps code structure 	! This is a huge update and took forever so I'll probably be taking a break for a few days + Added required includes to the files that had previously gotten them from other their parents * Lots of Makefile changes * Moved Room child definitions to resources/resources.hpp with all the other resources * Changed the declaration type for Object and Room childs to BEE::Object* and BEE::Room*

+++ bee/util/collision.hpp
@@ -11,9 +11,13 @@
 
 // Collision checking functions
 
+#include <SDL2/SDL.h>
+
+#include "real.hpp"
+
 class Line {
         public:
-                float x1, y1, x2, y2;
+                double x1, y1, x2, y2;
 };
 
 bool check_collision(SDL_Rect* a, SDL_Rect* b) {
@@ -44,23 +48,23 @@ bool check_collision(SDL_Rect* a, SDL_Rect* b) {
 
 	//return (SDL_HasIntersection(a, b) == SDL_TRUE) ? true : false;
 }
-bool check_collision_circle(float x1, float y1, float r1, float x2, float y2, float r2) {
+bool check_collision_circle(double x1, double y1, double r1, double x2, double y2, double r2) {
         if (distance(x1, y1, x2, y2) > r1+r2) {
                 return false;
         }
         return true;
 }
 bool check_collision_line(Line line1, Line line2) {
-        float m1 = (line1.y2-line1.y1) / (line1.x2-line1.x1); // Slope
-        float m2 = (line2.y2-line2.y1) / (line2.x2-line2.x1);
-        float b1 = line1.y1 - m1*line1.x1; // y-intercept
-        float b2 = line2.y1 - m2*line2.x1;
+        double m1 = (line1.y2-line1.y1) / (line1.x2-line1.x1); // Slope
+        double m2 = (line2.y2-line2.y1) / (line2.x2-line2.x1);
+        double b1 = line1.y1 - m1*line1.x1; // y-intercept
+        double b2 = line2.y1 - m2*line2.x1;
 
         if ((m1 == m2)&&(b1 == b2)) {
                 return true;
         }
 
-        float cx = (b2-b1) / (m2-m1); // Possible collision
+        double cx = (b2-b1) / (m2-m1); // Possible collision
 
         if ((cx >= line1.x1)&&(cx <= line1.x2)) {
                 if ((cx >= line2.x1)&&(cx <= line2.x2)) {
